twitter_geocoder/backend/api/requirements.txt
Django==4.2
djangorestframework==3.14.0
pytz==2022.1
django-cors-headers==3.14.0
python-decouple=3.8